# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: d4y API
  description: REST-API der d4y Git-native Runtime Platform. Read-only bezüglich des
    Sollzustands (ADR-0004); operative Aktionen sind transient (ADR-0013). Der Akteur
    wird über den Header X-Actor übermittelt.
  version: 0.1.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/reconcile:
    post:
      tags:
      - trigger-controller
      operationId: reconcile
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerRequest"
      responses:
        "200":
          description: OK
...
components:
  schemas:
    TriggerRequest:
      type: object
      properties:
        secrets:
          type: object
          additionalProperties:
            type: string
...